Use cfg_if to select between different stack_overflow handlers
No functional changes intended.
